# api/openapi/cloud-access-core.yaml
openapi: 3.1.0
info:
  title: Cloud Access Core API
  version: 0.9.0
  description: |
    REST surface for IAM, device onboarding, policy, and config center modules inside
    the 面试宝典 cloud-access-core service. JWT-protected admin endpoints use the IAM-issued
    access token; device endpoints use the device PASETO token. Deterministic error codes follow
    the scheme documented in docs/cloud-access-core-operations.md.
servers:
  - url: http://localhost:8080
    description: Default developer runtime (Makefile/Compose)
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceToken:
      type: http
      scheme: bearer
      bearerFormat: PASETO v4
  schemas:
...
    PairingRequest:
      type: object
      required: [platform, fingerprint]
      properties:
        platform:
          type: string
          example: windows
        fingerprint:
          type: string
        client_version:
          type: string
        display_name:
          type: string
        metadata:
          type: object
          additionalProperties: true
    PairingApprove:
      type: object
      required: [code]
      properties:
        code:
          type: string
        profile_name:
          type: string
          default: default
        trust_score:
          type: integer
          minimum: 0
          maximum: 100
    PairingClaim:
      type: object
      required: [code, token]
      properties:
        code:
          type: string
        token:
          type: string
    Heartbeat:
      type: object
      required: [device_id, status]
      properties:
        device_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [healthy, degraded, suspect]
        metrics:
          type: object
          additionalProperties: true
    RevokeDevice:
      type: object
      required: [device_id]
      properties:
        device_id:
          type: string
          format: uuid
        reason:
          type: string
    PolicyRule:
      type: object
      required: [action, resource, effect]
      properties:
        action:
          type: string
        resource:
          type: string
        effect:
          type: string
          enum: [allow, deny]
        roles:
          type: array
          items:
            type: string
        conditions:
          type: object
          additionalProperties: true
    PolicySet:
      type: object
      required: [tenant_id, rules]
      properties:
        tenant_id:
          type: string
          format: uuid
        rules:
          type: array
          items:
            $ref: '#/components/schemas/PolicyRule'
    CheckAccessRequest:
      type: object
      required: [tenant_id, action, resource]
      properties:
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
        action:
          type: string
        resource:
          type: string
    CheckAccessResponse:
      type: object
      properties:
        allowed:
          type: boolean
        reason:
          type: string
    QuotaRequest:
      type: object
      required: [user_id, resource, amount]
      properties:
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
        resource:
          type: string
        amount:
          type: integer
    QuotaResponse:
      type: object
      properties:
        allowed:
          type: boolean
        remaining:
          type: integer
        reset_at:
          type: string
          format: date-time
    GroupRequest:
      type: object
      required: [tenant_id, name]
      properties:
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        speech_quota:
          type: integer
        llm_quota:
          type: integer
        screenshot_quota:
          type: integer
        updated_by:
          type: string
          format: uuid
        overdraft_policy:
          type: object
          additionalProperties: true
    TenantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        tier:
          type: string
...
paths:
...
  /devices/pairing/request:
    post:
      summary: Start device pairing (anonymous client)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairingRequest'
      responses:
        '200':
          description: Pairing code issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  token:
                    type: string
                  expires_in:
                    type: integer
  /devices/pairing/approve:
    post:
      summary: Approve a device pairing request
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairingApprove'
      responses:
        '200':
          description: Device approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_id:
                    type: string
                  status:
                    type: string
                  config_version:
                    type: integer
  /devices/pairing/claim:
    post:
      summary: Device claims approved pairing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairingClaim'
      responses:
        '200':
          description: Encrypted profile bundle
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_id:
                    type: string
                  config_version:
                    type: integer
                  encrypted_profile:
                    type: string
  /devices/heartbeat:
    post:
      summary: Report device heartbeat + metrics
      security:
        - DeviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Heartbeat'
      responses:
        '202':
          description: Accepted for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
  /devices/revoke:
    post:
      summary: Revoke a device session
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeDevice'
      responses:
        '200':
          description: Device revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked:
                    type: boolean
  /policies:
    get:
      summary: List tenant policy rules
      security:
        - BearerToken: []
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Policy rules for the tenant
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/PolicyRule'
    put:
      summary: Replace tenant policy rules
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicySet'
      responses:
        '200':
          description: Updated rule set
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/PolicyRule'
  /policies/check:
    post:
      summary: Evaluate RBAC decision
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckAccessRequest'
      responses:
        '200':
          description: Decision payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckAccessResponse'
  /policies/evaluate:
    post:
      summary: Evaluate quota bucket and optionally deduct tokens
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotaRequest'
      responses:
        '200':
          description: Quota evaluation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaResponse'
  /groups:
    post:
      summary: Upsert a tenant group definition
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
      responses:
        '200':
          description: Persisted group model
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /tenants:
    post:
      summary: Create a new tenant
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...